Basic 4d vector implementation
Currently the implementation assumes that with compiler optimization on,
inlining of the functions will occur which will make the functions
(which return a struct) work
on the caller stack therefore will not issue a copy of the struct - this needs to
be checked, if it doesn't work, the functions should be changed to work
on the first pointer parameter.

The unit tests should be changed, current implementation is very basic
and uninformative. also it should have a separate build from the regular
tools.