Refactored matrix API.
Again removed the need for pointers for parameters on the matrix API,
for simple usage. The matrix struct is big, and this will probably be
the most performance hit, but currently, the performance is not the
focus and this can be changed later.

Remembered now to also change the API calls on the ReadMe of each
chapter.