Finished chapter 6
Chapter 6 introduced lighting, the implementation for the chapter
output, Helped to find a problem with color_as_u32 implementation -
there was no clamping of values, so in case of numbers bigger or smaller
than 1/0, the color values overflow, which result in unwanted colors.

Output can be found on answers_output/chapter6.